feat: rename accounts.ts to nimAccounts.ts
