+++ frontend/src/pages/modelServing/screens/projects/useNIMAccountConstants.ts
@@ -1,17 +1,17 @@
 import * as React from 'react';
 import { fetchNIMAccountConstants } from '~/pages/modelServing/screens/projects/nimUtils';
 
-type NIMAccountConstants =
-  | {
-      nimSecretName: string;
-      nimNGCSecretName: string;
-      nimConfigMapName: string;
-      templateName: string;
-    }
-  | undefined;
+type NIMAccountConstants = {
+  nimSecretName: string;
+  nimNGCSecretName: string;
+  nimConfigMapName: string;
+  templateName: string;
+};
 
-export const useNIMAccountConstants = (dashboardNamespace: string): NIMAccountConstants => {
-  const [constants, setConstants] = React.useState<NIMAccountConstants>(undefined);
+export const useNIMAccountConstants = (
+  dashboardNamespace: string,
+): NIMAccountConstants | undefined => {
+  const [constants, setConstants] = React.useState<NIMAccountConstants>();
 
   React.useEffect(() => {
     const fetchConstants = async () => {
